requirements.txt
# ── GUI
customtkinter>=5.2.0

# ── Hashing (fastest)
blake3>=0.3.0

# ── Compression
zstandard>=0.22.0
lz4>=4.3.0
brotli>=1.1.0
py7zr>=0.21.0
python-snappy>=0.7.1

# ── File type detection
python-magic>=0.4.27

# ── Image processing + perceptual hash
Pillow>=10.0.0
imagehash>=4.3.1
piexif>=1.1.3

# ── PDF metadata
pymupdf>=1.23.0
pypdf>=3.0.0

# ── Office metadata
python-docx>=1.0.0
openpyxl>=3.1.0
python-pptx>=0.6.21

# ── Audio/video metadata
mutagen>=1.47.0

# ── Cloud connectors
boto3>=1.34.0
google-api-python-client>=2.118.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
dropbox>=12.0.0

# ── Network / LAN P2P
zeroconf>=0.131.0

# ── Watch mode
watchdog>=4.0.0

# ── Encryption
cryptography>=42.0.0

# ── Storage analytics
plotly>=5.20.0

# ── YAML config
PyYAML>=6.0

# ── Scheduling
schedule>=1.2.0

# ── System info (drive type detection)
psutil>=5.9.0
